requirements.txt
discord.py==1.7.3
jikanpy==4.2.2
matplotlib==3.4.0
nest_asyncio==1.5.1
numpy==1.20.1
pymongo==4.0.2
Quart==0.15.1
requests==2.25.1
tracemoepy==4.1
youtube_dl==2021.5.16
